+++ akka-http-core/src/main/scala/akka/http/model/HttpForm.scala
@@ -1,6 +1,6 @@
 package akka.http.model
 
-import headers.`Content-Disposition`
+import akka.http.model.headers.`Content-Disposition`
 
 sealed trait HttpForm {
   type FieldType
